Minor fixes
This fixes the pad connections between P6 and P7, P9 and P10, P12 and P13 and the end of P15. Also it removes some redundant lines and adds a new connection between P4 and P5. Note that I have no idea how to fix the GND fill below P15 through.

Signed-off-by: Thomas Rohloff <v10lator@myway.de>